+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Added
+
+* `client::EventLogMultiplexer` (#266)
+
 ## 0.8.6 &ndash; 2022-07-18
 
 ## 0.8.5 &ndash; 2022-06-06